+++ src/remote.cpp
@@ -85,7 +85,7 @@ struct Packet23Data {
     int16_t grid_x;
     int16_t grid_y;
     uint8_t team;
-    uint8_t hits;
+    uint16_t hits;
     uint8_t speed;
     uint8_t shots;
     int16_t storage;
